Only show ipv4 address in preferences.
Previously the preferences dialog showed both ipv4 and ipv6 address if the interface had both.
This adjusts it to only show the ipv4 address, as sacn/multicast does not care about the ipv6 side.

It's likely that this change should end up in nicpicker as well, but I didn't want to duplicate the code until necessary.